#31 결제 기능 작성 완료(수정)
