+++ src/main/java/com/_3dhs/tnproject/payment/dto/TissueDTO.java
@@ -1,4 +1,4 @@
-package com._3dhs.tnproject.purchase.dto;
+package com._3dhs.tnproject.payment.dto;
 
 import lombok.Getter;
 import lombok.Setter;
