+++ src/main/resources/templates/common/testhub.html
@@ -21,7 +21,7 @@
 <hr>
 
 <h3>결제</h3>
-<button onclick="location.href='/purchase/purchase'">인기글 목록 조회</button>
+<button onclick="location.href='/payment/payment'">결제</button>
 <hr>
