+++ src/main/resources/templates/payment/payment.html
@@ -2,35 +2,17 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>purchase</title>
+    <title>payment</title>
     <link rel="stylesheet" href="/css/common.css">
     <!-- jQuery -->
     <script type="text/javascript" src="https://code.jquery.com/jquery-1.12.4.min.js" ></script>
-    <!-- iamport.payment.js -->
-<!--    <script type="text/javascript" src="https://cdn.iamport.kr/js/iamport.payment-1.2.0.js"></script>-->
     <script src="https://cdn.iamport.kr/v1/iamport.js"></script>
 
 </head>
 
 <body>
 <script>
-    //TODO  userId 받아오기
     memberId = `[[${loginUserId}]]`;
-    //memberId = `[[${member}]]`;
-    // var IMP = window.IMP;
-    // IMP.init("imp74535646");
-
-    // function requestPay() {
-    //     IMP.request_pay({
-    //     }, function (rsp) { // callback
-    //         $.ajax({
-    //             type: 'POST',
-    //             url: '/verify/' + rsp.imp_uid
-    //         }).done(function(data) {
-    //             if(rsp.paid_amount === data.response.amount){
-    //         });
-    //     });
-    // }
 
     function requestPay(tissuePrice) {
 
@@ -40,45 +22,26 @@
         IMP.request_pay({
             pg: "html5_inicis",
             pay_method: "card",
-            merchant_uid: "purchase" + new Date().getTime(),
+            merchant_uid: "payment" + new Date().getTime(),
             name: "티슈" + tissuePrice + "개",
             amount: tissuePrice,
             buyer_email: memberId,
             buyer_name: '프로필닉네임',
-            // buyer_tel: "010-1234-5678",
-            // buyer_addr: "서울특별시 강남구 삼성동",
-            // buyer_postcode: "123-456",
-
         }, function (rsp) {
             console.log(rsp);
             if (rsp.success) {
                 jQuery.ajax({
-                    url: "/purchaseSuccess",
+                    url: "/paymentSuccess",
                     method: "POST",
-                    // headers: {"Content-Type": "application/json"},
                     data: {
                         imp_uid: rsp.imp_uid,            // 결제 고유번호
                         merchant_uid: rsp.merchant_uid,  // 주문번호
                         buyer_name: rsp.buyer_name,
                         tissuePrice: tissuePrice,
                     },
-                    // jQuery.post("/purchase_success", {
-                    //     imp_uid: rsp.imp_uid,
-                    //     merchant_uid: rsp.merchant_uid
-                    // })
-                    // .done(function (data) {
-                    //
-                    //     // 고객사 서버 결제 API 성공시 로직
-                    //
-                    //     alert("성공");
-                    //     window.location.href = data.resultPath;
-                    // })
-                    // alert("성공");
-                    // window.location.href = "/purchase/purchase_success";
                     success : () => {
                         alert("성공");
-                        location.href = "/purchase/purchase_success";
-                        // window.location.href = "/payment/callback?imp_uid=" + rsp.imp_uid + "&merchant_uid=" + rsp.merchant_uid;
+                        location.href = "/payment/payment_success";
                     }
                 })
             } else {
